Add support for Kosovar/Albanian localization SHQIP
Hello,

I hope this message finds you well. I'm excited to contribute to this amazing project. Being a native speaker of Albanian language called (SHQIP - sq), I've noticed that our language wasn't supported in the current version of the project. To make this project more inclusive and accessible to the speakers of Kosova/Albania, I've decided to add localization support for my native language.

In this pull request, I have:

1. Added Shqip translation files to the  directory.
2. Updated the documentation to reflect this new language addition.

I believe that adding Shqip to the list of supported languages will make this project even more useful to a wider audience.
I look forward to your feedback and hope for the successful merger of this contribution. Please feel free to suggest any changes or improvements.

Thank you for considering this request.